move combine injector to core lib
